requirements.txt
django==5.2.1
djangorestframework==3.15.1
psycopg2-binary==2.9.9
celery==5.3.6
redis==5.0.3
python-dotenv==1.0.1
django-celery-beat>=2.7.0
aiogram>=3.3.0
aiogram-dialog>=2.1.0
aiohttp>=3.8.0
